Changed: move register on a top level
